+++ backend/application/reports/usecases/GetReportsUsecase.ts
@@ -1,5 +1,5 @@
-import { IReportsRepository } from '../../domain/repositories/IReportsRepository';
-import { Reports } from '../../domain/entities/Reports';
+import { Reports } from '@/backend/domain/entities/Report';
+import { IReportsRepository } from '@/backend/domain/repositories/IReportsRepository';
 
 export class GetReportsUsecase {
   constructor(private readonly reportsRepository: IReportsRepository) {}
